refactor HALValue: removed as<type> replaced it with to<type> and asRaw<type> also did refactor of whole HALValue to make it consistent + added new WiFiConnection manager

+++ src/DALHAL/Devices/GeneralOutputs/SinglePulseOutput/DALHAL_SinglePulseOutput.cpp
@@ -75,7 +75,7 @@ namespace DALHAL {
         if (val.getType() == HALValue::Type::TEST) { /*printf("\nSinglePulseOutput::write TEST\n");*/ return HALOperationResult::Success; }// test write to check feature
         if (val.isNaN()) return HALOperationResult::WriteValueNaN;
 
-        uint32_t t = val;
+        uint32_t t = val.toUInt();
         if (t != 0) {// only change if not zero
             
 #if HAS_REACTIVE(SINGLE_PULSE_OUTPUT, VALUE_CHANGE)
